+++ ms-route/src/main/java/com/thelak/route/payments/models/CertificateModel.java
@@ -16,7 +16,7 @@ public class CertificateModel {
 
     Long id;
 
-    Integer months;
+    Long months;
 
     String name;
